+++ doc/changelog.rst
@@ -25,6 +25,7 @@ MJX
 - Added sphere-cylinder and sphere-ellipsoid collision functions (:github:issue:`2126`).
 - Fixed a bug with frictionloss constraints.
 - Added ``TENDONPOS`` sensor.
+- Fixed a bug with the computation of tangential contact forces in ``_decode_pyramid``.
 
 Bug fixes
 ^^^^^^^^^